+++ models/job/RemoveMaterialDirJob.php
@@ -11,13 +11,23 @@ class RemoveMaterialDirJob extends BaseObject implements JobInterface
 {
     public $materials;
 
+    /**
+     * Deleting work directory materials
+     * @param \yii\queue\Queue $queue
+     * @return bool
+     */
     public function execute($queue)
     {
         foreach ($this->materials as $material) {
 
-            $dir = \Yii::$app->getBasePath().\Yii::$app->params['PathToAttachments'].$material->dir;
+            if ($material->dir) {
 
-            Material::removeDirectory($dir);
+                $dir = \Yii::$app->getBasePath().\Yii::$app->params['PathToAttachments'].$material->dir;
+
+                if (file_exists($dir)) {
+                    Material::removeDirectory($dir);
+                }
+            }
         }
 
         return true;
